refactor: Rename `lib/plugins/aws/lib/getServiceState.js`
