+++ lib/plugins/aws/lib/get-service-state.js
