+++ lib/plugins/aws/package/compile/events/apiGateway/index.js
@@ -260,7 +260,7 @@ class AwsCompileApigEvents {
 
       // TODO should be removed once AWS fixes the CloudFormation problems using a separate Stage
       'after:deploy:deploy': async () => {
-        const getServiceState = require('../../../../lib/getServiceState').getServiceState;
+        const getServiceState = require('../../../../lib/get-service-state').getServiceState;
 
         const state = getServiceState.call(this);
         const updateStage = require('./lib/hack/updateStage').updateStage;
