+++ test/unit/lib/plugins/aws/lib/get-service-state.test.js
@@ -5,7 +5,7 @@ const chai = require('chai');
 const sinon = require('sinon');
 const Serverless = require('../../../../../../lib/serverless');
 const AwsProvider = require('../../../../../../lib/plugins/aws/provider');
-const getServiceState = require('../../../../../../lib/plugins/aws/lib/getServiceState');
+const getServiceState = require('../../../../../../lib/plugins/aws/lib/get-service-state');
 
 const expect = chai.expect;
 chai.use(require('sinon-chai'));
