+++ test/unit/lib/plugins/aws/package/compile/events/apiGateway/index.test.js
@@ -6,7 +6,7 @@ const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider
 const AwsCompileApigEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/apiGateway/index');
 const Serverless = require('../../../../../../../../../lib/serverless');
 const validate = require('../../../../../../../../../lib/plugins/aws/lib/validate');
-const getServiceState = require('../../../../../../../../../lib/plugins/aws/lib/getServiceState');
+const getServiceState = require('../../../../../../../../../lib/plugins/aws/lib/get-service-state');
 const updateStage = require('../../../../../../../../../lib/plugins/aws/package/compile/events/apiGateway/lib/hack/updateStage');
 const disassociateUsagePlan = require('../../../../../../../../../lib/plugins/aws/package/compile/events/apiGateway/lib/hack/disassociateUsagePlan');
